+++ features/Posts/api.ts
@@ -18,7 +18,7 @@ export function getPostBySlug(slug: string, fields: string[] = []) {
   const { data, content } = matter(fileContents);
 
   type Items = {
-    [key: string]: string;
+    [key: string]: string | Array<string>;
   };
 
   const items: Items = {};