+++ features/Posts/components/post/Post.tsx
@@ -6,6 +6,9 @@ import PostHeader, { PostHeaderType } from "./post-header/PostHeader";
 import PostNavigation, {
   PostNavigationType,
 } from "./post-navigation/PostNavigation";
+import PostRelatedArticles, {
+  PostRelatedArticlesType,
+} from "./post-related-articles/PostRelatedArticles";
 
 interface IPostProps {
   children: ReactNode;
@@ -15,6 +18,7 @@ interface IPostComposition {
   Header: PostHeaderType;
   Content: PostContentType;
   Navigation: PostNavigationType;
+  RelatedArticles: PostRelatedArticlesType;
 }
 
 type PostProps = FunctionComponent<IPostProps> & IPostComposition;
@@ -35,5 +39,6 @@ const Post: PostProps = ({ children }) => {
 Post.Header = PostHeader;
 Post.Content = PostContent;
 Post.Navigation = PostNavigation;
+Post.RelatedArticles = PostRelatedArticles;
 
 export default Post;