+++ pages/blog/[slug].tsx
@@ -4,7 +4,7 @@ import ErrorPage from "next/error";
 
 import { PrimaryLayout } from "features/Layout";
 
-import { POST_ALL_FIELDS, Post } from "features/Posts";
+import { POST_ALL_FIELDS, Post, POST_HEADER_FIELDS } from "features/Posts";
 
 import {
   getAllPostsByDate,
@@ -17,15 +17,16 @@ import markdownToHtml from "core/utilities/markdownToHtml";
 
 import { SITE_IMAGE } from "core/constants";
 
-import type { IPost, PostAnchor } from "features/Posts/types";
+import type { IPost, PostAnchor, PostSummary } from "features/Posts/types";
 
 type Props = {
   post: IPost;
   previous: PostAnchor | null;
   next: PostAnchor | null;
+  related: Array<PostSummary>;
 };
 
-export default function PostPage({ post, previous, next }: Props) {
+export default function PostPage({ post, previous, next, related }: Props) {
   const router = useRouter();
   if (!router.isFallback && !post?.slug) {
     return <ErrorPage statusCode={404} />;
@@ -47,10 +48,16 @@ export default function PostPage({ post, previous, next }: Props) {
           tags={post.tags}
         />
         <Post.Content content={post.content} />
-        <Post.Navigation
-          previous={previous && { title: previous.title, href: previous.slug }}
-          next={next && { title: next.title, href: next.slug }}
-        />
+        {related.length > 0 ? (
+          <Post.RelatedArticles relatedArticles={related} />
+        ) : (
+          <Post.Navigation
+            previous={
+              previous && { title: previous.title, href: previous.slug }
+            }
+            next={next && { title: next.title, href: next.slug }}
+          />
+        )}
       </Post>
     </PrimaryLayout>
   );
@@ -65,11 +72,16 @@ type Params = {
 export async function getStaticProps({ params }: Params) {
   const post = getPostBySlug(params.slug, POST_ALL_FIELDS);
 
-  const content = await markdownToHtml(post.content || "");
+  const content = await markdownToHtml((post.content as string) || "");
 
   const previous = getPreviousPost(params.slug);
-
   const next = getNextPost(params.slug);
+
+  const relatedSlugs = (post.related as Array<string>) || [];
+
+  const related = relatedSlugs
+    .slice(0, 2)
+    .map((rs) => getPostBySlug(rs, POST_HEADER_FIELDS));
 
   return {
     props: {
@@ -79,6 +91,7 @@ export async function getStaticProps({ params }: Params) {
       },
       previous,
       next,
+      related,
     },
   };
 }